# app/Models/Level.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Concerns\HasUuids;
use Illuminate\Database\Eloquent\Model;

class Level extends Model
{
	use HasUuids;

	public $timestamps = false;

	protected $fillable = [
		'name',
		'description'
	];
}